fixed offset calc and relocated stuff.
