+++ tools/meson.build
@@ -1,8 +1,6 @@
-libelf = dependency('libelf', required: true)
-
 executable(
   'tux-run',
-  sources: ['tux_run.c', 'perf_support.c'],
-  dependencies: [libtux_dep, libpal_lfi_dep, argp, libelf],
+  sources: ['tux_run.c'],
+  dependencies: [libtux_dep, libpal_lfi_dep, argp],
   install: true,
 )
